requirements.txt
langchain
openai
tiktoken
pinecone-client
googletrans==3.1.0a0
pypdf

